organized files and created a reset css
