+++ src/Components/Header.js
@@ -1,4 +1,4 @@
-import react from "react";
+import React from "react";
 
 function Header () {
     return (
