+++ src/Components/Navigation.js
@@ -1,5 +1,5 @@
 import React from "react";
-import {information} from "../Information.js";
+import {information} from "../information.js";
 import {Link} from "react-router-dom";
 
 function Navigation (props) {
@@ -8,7 +8,7 @@ function Navigation (props) {
     return (
         <div className="navigation" onClick={() => routeChange()} >
             {information.Routes.map((route, i) => {
-                return <Link to={route == "Home"? "/": route.toLowerCase()} key={i}
+                return <Link to={route === "Home"? "/": route.toLowerCase()} key={i}
                     className="navLink" >{route}</Link>
             })}
         </div>
