+++ src/index.js
@@ -1,8 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import './index.css';
 import App from './App';
 import {BrowserRouter} from "react-router-dom";
+
+import "./styles/reset.css";
+import './styles/index.css';
+import "./styles/App.css";
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
